% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceBlogTools.R
\name{create_index_qmd_file}
\alias{create_index_qmd_file}
\title{Automate steps for contributing blogposts to the Epiverse-Trace blog.}
\usage{
create_index_qmd_file(
  post_folder_name,
  posts_dir = file.path("./posts"),
  title = NULL,
  subtitle = NULL,
  author = NULL,
  open_file = FALSE
)
}
\arguments{
\item{post_folder_name}{A character string with the name of the folder to
store the new post. Note that the post_folder_name will be used as the
branch name so it must be valid R object name (A hyphenated name is
preferred, for example, "benchmarking-dt" but any valid R name will work).}

\item{posts_dir}{A character string with the path to the posts directory}

\item{title}{A character string with the title of the post}

\item{subtitle}{A character string with the subtitle of the post}

\item{author}{A character string with the author of the post}

\item{open_file}{A logical value indicating whether to open the index.qmd
file after it is created}
}
\value{
A new index.qmd file
}
\description{
This is the main function that automates the steps outlined in the
\href{https://github.com/epiverse-trace/epiverse-trace.github.io/blob/main/.github/CONTRIBUTING.md#step-by-step}{contributing guide}). #nolint
It creates a git branch using the name of the post folder, adds the
directory "./posts/<post_folder_name>/ folder, and adds a boiler plate
\code{index.qmd} file.
}
\examples{
\dontrun{
create_index_qmd_file("benchmarking-dt")
}
}
