% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraceBlogTools.R
\name{create_post_git_branch}
\alias{create_post_git_branch}
\title{Create a new git branch}
\usage{
create_post_git_branch(post_folder_name)
}
\arguments{
\item{post_folder_name}{A character string with the name of the folder to
store the new post. Note that the post_folder_name will be used as the
branch name so it must be valid R object name (A hyphenated name is
preferred, for example, "benchmarking-dt" but any valid R name will work).}
}
\value{
A new git branch
}
\description{
Create a new git branch
}
\examples{
\dontrun{
create_post_git_branch("benchmarking-dt")
}
}
