% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{TRACEBlogUtils-package}
\alias{TRACEBlogUtils}
\alias{TRACEBlogUtils-package}
\title{TRACEBlogUtils: Utility Functions For Contributing Blog Posts to the Epiverse-TRACE Blog}
\description{
Makes contributing blogposts to the Epiverse-Trace blog as seamless as possible in a manner similar to the usethis package.
}
\author{
\strong{Maintainer}: James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [copyright holder]

}
\keyword{internal}
